requirements.txt
py-clob-client
python-dotenv
pandas
gspread
gspread-dataframe
sortedcontainers
black
eth-account
eth-utils
poly_eip712_structs
py_order_utils
pytest
python-dotenv
requests
websockets
cryptography
asyncio
requests
google-auth
web3
scipy
ruff
ty
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp